Zmiana 2x tabel small int na int oraz dl pola navchar

+++ SalonSamochodowy/SalonSamochodowy/Produkt/Tables/Klient.sql
@@ -5,12 +5,14 @@
     [KlientImieDrugie]   NVARCHAR (20) NOT NULL,
     [KlientNazwisko]     NVARCHAR (35) NOT NULL,
     [DataDodania]        DATE          NOT NULL,
-    [Telefon]            SMALLINT      NOT NULL,
+    [Telefon]            INT           NOT NULL,
     [EmailId]            NVARCHAR (45) NOT NULL,
     [DowodOsobistyId]    INT           NULL,
     PRIMARY KEY CLUSTERED ([KlientId] ASC),
     FOREIGN KEY ([DowodOsobistyId]) REFERENCES [Produkt].[Do] ([DoId]) ON DELETE SET NULL
 );
+
+
